move PageNotFoundException class to org.wso2.carbon.uis.api.exception
